Use chdir option to exec instead of FileUtils.cd
Using `FileUtils.cd` (an alias for `Dir.chdir`) triggers a warning
message when Subprocess is called with the cwd arg from inside another
chdir block:
https://github.com/ruby/ruby/blob/v2_4_1/dir.c#L1033-L1034

In #29, I tried fixing this by using the block form of `Dir.chdir` but
that caused many people to see errors of the form:
```
<elided>/subprocess-1.3.7/lib/subprocess.rb:280:in `exec': Too many open files - getcwd (Errno::EMFILE)
```

Instead of calling `chdir` in Ruby, I think it makes more sense to pass
`chdir` as an argument to exec.

I still have to call `Dir.chdir` around a `preexec_fn` so that it
continues to operate in the new working directory. We don't depend on
that behavior anywhere at Stripe but it's possible someone else
does. I suspect the bug was some interaction between `Dir.chdir` and
`exec` so this feels safe.