Add support for partial reads from Subprocess#communicate
It can be useful to read partial output from a long running process before
it completes (e.g. to provide feedback to the user about the progress).
This adds the ability to pass a block to `communicate` that gets called
with data read from stdout/stderr as its read instead of returning it.