Do not hang in communicate if spawned process forks
Currently we only exit the `communicate` loop when we get an EOF from
stdout and stderr. If the child process forks before exiting, the fork
will inherit those file descriptors and `communicate` will hang until
the child exits. I believe it's preferable to drain stdout/stderr after
the process exits and then break out of the `communicate` loop.

Note that Open3 similarly hangs waiting for EOF. There is a risk that
existing software depends on this behavior and that output could be
non-deterministic if the forked grandchild is writing to stdout/stderr.